requirements.txt
pandas
numpy
seaborn
matplotlib
python-dotenv
certifi
scikit-learn
dill
xgboost
catboost
ipykernel
pymongo
pymongo[srv]==3.12
pyaml
mlflow